Create DC_ScanHTML_JS_Tree.ps1
Say it with me, Do Not Mess With The RegEx.  It came from this brilliant site because my attempt was MESSY.

This creates a JS and a CSS file in the directory specified in the script.  Seems to work fine on all tested browsers, but YMMV. 
YOLO. 
IANAL. 

If you do something awesome with this, let me know.